requirements.txt
Flask
requests
beautifulsoup4
lxml
cloudinary
python-dotenv

